STREAMLIT/requirements.txt
streamlit
snowflake-connector-python
pandas
matplotlib
plotly
nltk
scikit-learn
requests
python-dotenv
openpyxl               
folium 
pydeck       
geopy 
wikipedia-api==0.5.8
geopy==2.3.0
plotly==5.11.0
streamlit==1.22.0
pandas==1.5.2
requests==2.28.1